fix(tech-debt): Split AgentModals into separate files (Issue #114)
Split emdx/ui/agent_modals.py (879 lines, 9 classes) into separate modules:
- run_agent_modal.py: RunAgentModal class
- agent_history_modal.py: AgentHistoryModal class
- create_edit_agent_modal.py: CreateAgentModal and EditAgentModal classes
- delete_agent_modal.py: DeleteAgentModal class
- agent_selection_modal.py: AgentListItem, AgentListWidget, AgentDetailPanel, AgentSelectionModal

The original agent_modals.py now serves as a re-export module for backwards compatibility.
